repositionating test files
